refactor: PresetDropdownItemを独立したファイルに分離
- PresetDropdownItemを別ファイルへ抽出し、コンポーネントの責務をさらに明確化
- 既存の購読最適化（個別インデックスの購読）を維持
- PresetDropdownからのインポート形式に修正
